Extract `useBucketChangeOverview`
